+++ testsuite/Main.hs
@@ -10,88 +10,89 @@ import           System.FilePath  ((</>),(<.>))
 import qualified System.IO.Unsafe as Unsafe
 
 data BuildTarget
-  = VHDL | SystemVerilog
+  = VHDL | Verilog | Both
   deriving Show
 
 main :: IO ()
 main =
   defaultMain $ testGroup "tests"
     [ testGroup "examples"
-      [runTest "examples"             VHDL [] "ALU"          (Just ("topEntity",False))
+      [runTest "examples"             Both [] "ALU"          (Just ("topEntity",False))
       ,runTest "examples"             VHDL [] "Blinker"      (Just ("topEntity_0",False))
-      ,runTest "examples"             VHDL [] "BlockRamTest" (Just ("topEntity",False))
-      ,runTest "examples"             VHDL [] "Calculator"   (Just ("testbench",True ))
-      ,runTest "examples"             VHDL [] "CochleaPlus"  (Just ("topEntity",False))
-      ,runTest "examples"             VHDL [] "FIR"          (Just ("testbench",True ))
-      ,runTest "examples"             VHDL [] "Fifo"         (Just ("topEntity",False))
-      ,runTest "examples"             VHDL [] "MAC"          (Just ("testbench",True))
-      ,runTest "examples"             VHDL [] "MatrixVect"   (Just ("testbench",True))
-      ,runTest "examples"             VHDL [] "Queens"       (Just ("topEntity",False))
-      ,runTest "examples"             VHDL [] "Reducer"      (Just ("topEntity",False))
-      ,runTest "examples"             VHDL [] "Sprockell"    (Just ("topEntity",False))
-      ,runTest "examples"             VHDL [] "Windows"      (Just ("topEntity",False))
-      ,runTest ("examples" </> "i2c") VHDL [] "I2C"          (Just ("topEntity",False))
+      ,runTest "examples"             Verilog [] "Blinker_v" (Just ("topEntity_0",False))
+      ,runTest "examples"             Both [] "BlockRamTest" (Just ("topEntity",False))
+      ,runTest "examples"             Both [] "Calculator"   (Just ("testbench",True ))
+      ,runTest "examples"             Both [] "CochleaPlus"  (Just ("topEntity",False))
+      ,runTest "examples"             Both [] "FIR"          (Just ("testbench",True ))
+      ,runTest "examples"             Both [] "Fifo"         (Just ("topEntity",False))
+      ,runTest "examples"             Both [] "MAC"          (Just ("testbench",True))
+      ,runTest "examples"             Both [] "MatrixVect"   (Just ("testbench",True))
+      ,runTest "examples"             Both [] "Queens"       (Just ("topEntity",False))
+      ,runTest "examples"             Both [] "Reducer"      (Just ("topEntity",False))
+      ,runTest "examples"             Both [] "Sprockell"    (Just ("topEntity",False))
+      ,runTest "examples"             Both [] "Windows"      (Just ("topEntity",False))
+      ,runTest ("examples" </> "i2c") Both [] "I2C"          (Just ("topEntity",False))
       ]
     , testGroup "unit-tests"
         [ testGroup "Basic"
-            [ runTest ("tests" </> "shouldwork" </> "Basic") VHDL [] "ClassOps" (Just ("topEntity",False))
-            , runTest ("tests" </> "shouldwork" </> "Basic") VHDL [] "IrrefError" (Just ("topEntity",False))
-            , runTest ("tests" </> "shouldwork" </> "Basic") VHDL [] "NestedPrimitives" (Just ("topEntity",False))
-            , runTest ("tests" </> "shouldwork" </> "Basic") VHDL [] "NestedPrimitives2" (Just ("topEntity",False))
-            , runTest ("tests" </> "shouldwork" </> "Basic") VHDL [] "PatError" (Just ("topEntity",False))
-            , runTest ("tests" </> "shouldwork" </> "Basic") VHDL [] "RecordSumOfProducts" (Just ("topEntity",False))
-            , runTest ("tests" </> "shouldwork" </> "Basic") VHDL [] "Shift" (Just ("topEntity",False))
-            , runTest ("tests" </> "shouldwork" </> "Basic") VHDL [] "SimpleConstructor" (Just ("topEntity",False))
-            , runTest ("tests" </> "shouldwork" </> "Basic") VHDL [] "TagToEnum" (Just ("topEntity",False))
-            , runTest ("tests" </> "shouldwork" </> "Basic") VHDL [] "TwoFunctions" (Just ("topEntity",False))
+            [ runTest ("tests" </> "shouldwork" </> "Basic") Both [] "ClassOps" (Just ("topEntity",False))
+            , runTest ("tests" </> "shouldwork" </> "Basic") Both [] "IrrefError" (Just ("topEntity",False))
+            , runTest ("tests" </> "shouldwork" </> "Basic") Both [] "NestedPrimitives" (Just ("topEntity",False))
+            , runTest ("tests" </> "shouldwork" </> "Basic") Both [] "NestedPrimitives2" (Just ("topEntity",False))
+            , runTest ("tests" </> "shouldwork" </> "Basic") Both [] "PatError" (Just ("topEntity",False))
+            , runTest ("tests" </> "shouldwork" </> "Basic") Both [] "RecordSumOfProducts" (Just ("topEntity",False))
+            , runTest ("tests" </> "shouldwork" </> "Basic") Both [] "Shift" (Just ("topEntity",False))
+            , runTest ("tests" </> "shouldwork" </> "Basic") Both [] "SimpleConstructor" (Just ("topEntity",False))
+            , runTest ("tests" </> "shouldwork" </> "Basic") Both [] "TagToEnum" (Just ("topEntity",False))
+            , runTest ("tests" </> "shouldwork" </> "Basic") Both [] "TwoFunctions" (Just ("topEntity",False))
             ]
         , testGroup "BitVector"
-            [ runTest ("tests" </> "shouldwork" </> "BitVector") VHDL [] "Box" (Just ("topEntity",False))
+            [ runTest ("tests" </> "shouldwork" </> "BitVector") Both [] "Box" (Just ("topEntity",False))
             ]
         , testGroup "BoxedFunctions"
-            [ runTest ("tests" </> "shouldwork" </> "BoxedFunctions") VHDL [] "DeadRecursiveBoxed" (Just ("topEntity",False))
+            [ runTest ("tests" </> "shouldwork" </> "BoxedFunctions") Both [] "DeadRecursiveBoxed" (Just ("topEntity",False))
             ]
         , testGroup "CSignal"
-            [ runTest ("tests" </> "shouldwork" </> "CSignal") VHDL [] "CBlockRamTest" (Just ("topEntity",False))
-            , runTest ("tests" </> "shouldwork" </> "CSignal") VHDL [] "MAC" (Just ("topEntity",False))
+            [ runTest ("tests" </> "shouldwork" </> "CSignal") Both [] "CBlockRamTest" (Just ("topEntity",False))
+            , runTest ("tests" </> "shouldwork" </> "CSignal") Both [] "MAC" (Just ("topEntity",False))
             ]
         , testGroup "Fixed"
-            [ runTest ("tests" </> "shouldwork" </> "Fixed") VHDL [] "SFixedTest" (Just ("testbench",True))
+            [ runTest ("tests" </> "shouldwork" </> "Fixed") Both [] "SFixedTest" (Just ("testbench",True))
             ]
         , testGroup "Numbers"
-            [ runTest ("tests" </> "shouldwork" </> "Numbers") VHDL [] "Resize" (Just ("testbench",True))
-            , runTest ("tests" </> "shouldwork" </> "Numbers") VHDL [] "Resize2" (Just ("testbench",True))
-            , runTest ("tests" </> "shouldwork" </> "Numbers") VHDL [] "SatMult" (Just ("topEntity",False))
+            [ runTest ("tests" </> "shouldwork" </> "Numbers") Both [] "Resize" (Just ("testbench",True))
+            , runTest ("tests" </> "shouldwork" </> "Numbers") Both [] "Resize2" (Just ("testbench",True))
+            , runTest ("tests" </> "shouldwork" </> "Numbers") Both [] "SatMult" (Just ("topEntity",False))
             ]
         , testGroup "Polymorphism"
-            [ runTest ("tests" </> "shouldwork" </> "Polymorphism") VHDL [] "ExistentialBoxed" (Just ("topEntity",False))
-            , runTest ("tests" </> "shouldwork" </> "Polymorphism") VHDL [] "LocalPoly" (Just ("topEntity",False))
+            [ runTest ("tests" </> "shouldwork" </> "Polymorphism") Both [] "ExistentialBoxed" (Just ("topEntity",False))
+            , runTest ("tests" </> "shouldwork" </> "Polymorphism") Both [] "LocalPoly" (Just ("topEntity",False))
             ]
         , testGroup "Signal"
-            [ runTest ("tests" </> "shouldwork" </> "Signal") VHDL [] "AlwaysHigh" (Just ("topEntity",False))
-            , runTest ("tests" </> "shouldwork" </> "Signal") VHDL [] "BlockRamTest" (Just ("topEntity",False))
-            , runTest ("tests" </> "shouldwork" </> "Signal") VHDL [] "MAC" (Just ("topEntity",False))
-            , runTest ("tests" </> "shouldwork" </> "Signal") VHDL [] "SigP" (Just ("topEntity",False))
+            [ runTest ("tests" </> "shouldwork" </> "Signal") Both [] "AlwaysHigh" (Just ("topEntity",False))
+            , runTest ("tests" </> "shouldwork" </> "Signal") Both [] "BlockRamTest" (Just ("topEntity",False))
+            , runTest ("tests" </> "shouldwork" </> "Signal") Both [] "MAC" (Just ("topEntity",False))
+            , runTest ("tests" </> "shouldwork" </> "Signal") Both [] "SigP" (Just ("topEntity",False))
             ]
         , testGroup "Testbench"
-            [ runTest ("tests" </> "shouldwork" </> "Testbench") VHDL ["-clash-inline-limit=0"] "TB" (Just ("testbench",True))
-            , runTest ("tests" </> "shouldwork" </> "Testbench") VHDL [] "SyncTB" (Just ("testbench",True))
+            [ runTest ("tests" </> "shouldwork" </> "Testbench") Both ["-clash-inline-limit=0"] "TB" (Just ("testbench",True))
+            , runTest ("tests" </> "shouldwork" </> "Testbench") Both [] "SyncTB" (Just ("testbench",True))
             ]
         , testGroup "Vector"
-            [ runTest ("tests" </> "shouldwork" </> "Vector") VHDL [] "EnumTypes" (Just ("topEntity",False))
-            , runTest ("tests" </> "shouldwork" </> "Vector") VHDL [] "HOClock" (Just ("topEntity",False))
-            , runTest ("tests" </> "shouldwork" </> "Vector") VHDL [] "HOCon" (Just ("topEntity",False))
-            , runTest ("tests" </> "shouldwork" </> "Vector") VHDL [] "HOPrim" (Just ("topEntity",False))
-            , runTest ("tests" </> "shouldwork" </> "Vector") VHDL [] "PatHOCon" (Just ("topEntity",False))
-            , runTest ("tests" </> "shouldwork" </> "Vector") VHDL [] "Split" (Just ("topEntity",False))
-            , runTest ("tests" </> "shouldwork" </> "Vector") VHDL [] "VACC" (Just ("topEntity",False))
-            , runTest ("tests" </> "shouldwork" </> "Vector") VHDL [] "VIndex" (Just ("topEntity",False))
-            , runTest ("tests" </> "shouldwork" </> "Vector") VHDL [] "VMapAccum" (Just ("topEntity",False))
-            , runTest ("tests" </> "shouldwork" </> "Vector") VHDL [] "VReplace" (Just ("topEntity",False))
-            , runTest ("tests" </> "shouldwork" </> "Vector") VHDL [] "VScan" (Just ("topEntity",False))
-            , runTest ("tests" </> "shouldwork" </> "Vector") VHDL [] "VZip" (Just ("topEntity",False))
-            , runTest ("tests" </> "shouldwork" </> "Vector") VHDL [] "VecConst" (Just ("topEntity",False))
-            , runTest ("tests" </> "shouldwork" </> "Vector") VHDL [] "VecOfSum" (Just ("topEntity",False))
+            [ runTest ("tests" </> "shouldwork" </> "Vector") Both [] "EnumTypes" (Just ("topEntity",False))
+            , runTest ("tests" </> "shouldwork" </> "Vector") Both [] "HOClock" (Just ("topEntity",False))
+            , runTest ("tests" </> "shouldwork" </> "Vector") Both [] "HOCon" (Just ("topEntity",False))
+            , runTest ("tests" </> "shouldwork" </> "Vector") Both [] "HOPrim" (Just ("topEntity",False))
+            , runTest ("tests" </> "shouldwork" </> "Vector") Both [] "PatHOCon" (Just ("topEntity",False))
+            , runTest ("tests" </> "shouldwork" </> "Vector") Both [] "Split" (Just ("topEntity",False))
+            , runTest ("tests" </> "shouldwork" </> "Vector") Both [] "VACC" (Just ("topEntity",False))
+            , runTest ("tests" </> "shouldwork" </> "Vector") Both [] "VIndex" (Just ("topEntity",False))
+            , runTest ("tests" </> "shouldwork" </> "Vector") Both [] "VMapAccum" (Just ("topEntity",False))
+            , runTest ("tests" </> "shouldwork" </> "Vector") Both [] "VReplace" (Just ("topEntity",False))
+            , runTest ("tests" </> "shouldwork" </> "Vector") Both [] "VScan" (Just ("topEntity",False))
+            , runTest ("tests" </> "shouldwork" </> "Vector") Both [] "VZip" (Just ("topEntity",False))
+            , runTest ("tests" </> "shouldwork" </> "Vector") Both [] "VecConst" (Just ("topEntity",False))
+            , runTest ("tests" </> "shouldwork" </> "Vector") Both [] "VecOfSum" (Just ("topEntity",False))
             ]
         ]
     ]
@@ -102,7 +103,7 @@ clashHDL t env extraArgs modName =
               "cabal"
               (concat [["exec","clash","--"
                        ,case t of { VHDL -> "--vhdl"
-                                  ; SystemVerilog -> "--systemverilog"
+                                  ; Verilog -> "--verilog"
                                   }
                        ]
                       ,extraArgs
@@ -125,6 +126,18 @@ ghdlMake entity env = testProgram "GHDL (make)" "ghdl" ["-m","--workdir=work",en
 
 ghdlSim :: FilePath -> TestTree
 ghdlSim env = testProgram "GHDL (sim)" "ghdl" ["-r","testbench","--assert-level=error"] (Just env) False
+
+iverilog :: FilePath -> String -> TestTree
+iverilog dir entity = withResource (return dir) (const (return ()))
+    (\d -> testProgram "iverilog" "iverilog" ("-g2005":"-s":entity:"-o":entity:verilogFiles d) (Just dir) False)
+  where
+    verilogFiles :: IO FilePath -> [FilePath]
+    verilogFiles d =  Unsafe.unsafePerformIO
+                   $  filter (List.isSuffixOf "v")
+                  <$> (Directory.getDirectoryContents =<< d)
+
+vvp :: FilePath -> TestTree
+vvp env = testProgram "vvp" "vvp" ["-N","testbench"] (Just env) False
 
 runTest :: FilePath
         -> BuildTarget
@@ -149,4 +162,23 @@ runTest env VHDL extraArgs modName entNameM = withResource aquire release (const
                                    , ghdlMake entName modDir
                                    ] ++ if doRun then [ghdlSim modDir] else []
 
-runTest env SystemVerilog extraArgs modName _ = clashHDL SystemVerilog env extraArgs modName
+runTest env Verilog extraArgs modName entNameM =
+    withResource (return ()) release (const grp)
+  where
+    verilogDir = env </> "verilog"
+    modDir     = verilogDir </> modName
+    release _  = Directory.removeDirectoryRecursive verilogDir
+
+
+    grp        = testGroup modName
+                    ( clashHDL Verilog env extraArgs modName
+                    : maybe [] doMakeAndRun entNameM
+                    )
+
+    doMakeAndRun (entName,doRun) = [ iverilog modDir entName
+                                   ] ++ if doRun then [vvp modDir] else []
+
+runTest env Both extraArgs modName entNameM = testGroup "VHDL & Verilog"
+  [ runTest env VHDL extraArgs modName entNameM
+  , runTest env Verilog extraArgs modName entNameM
+  ]